logout redirects to home

+++ lambdafit/src/components/logout.js
@@ -1,15 +1,26 @@
-import React from 'react';
+import { useEffect } from 'react';
 
-export default function Logout() {
-	const onLogOut = () => {
-		localStorage.clear();
-	};
+export default function Logout(props) {
+	// const onLogOut = () => {
+	// 	localStorage.clear();
+	// };
 
-	return (
-		<div>
-			<button type="submit" onClick={onLogOut}>
-				Log Out
-			</button>
-		</div>
-	);
+	// return (
+	// 	<div>
+	// 		<button type="submit" onClick={onLogOut}>
+	// 			Log Out
+	// 		</button>
+	// 	</div>
+	// );
+
+	useEffect(() => {
+		if (props.isLoggedIn) {
+			localStorage.clear();
+			window.location = '/';
+		} else {
+			window.location = '/';
+		}
+	});
+
+	return null;
 }
